Versão final que foi entregue ao professor. Coeficiente soften finalizado Corrigido calculo da refração

+++ trunk/RayTracing/Src/OmniLight.h
@@ -7,5 +7,5 @@ public:
 	OmniLight( void );
 	OmniLight( const Vector3D &aPosition, const ColorRGBf &aColor = ColorRGBf(1,1,1));
 
-	virtual double evaluateIlumination( const Vector3D& aPoint, const Scene* aScene, double aTiny);
+	virtual float evaluateIlumination( const Vector3D& aPoint, const Scene* aScene, double aTiny);
 };
